fix: shift methods to main page
